Modify: ID 타입을 ERD 기준에 맞춰 Long에서 Integer로 변경

+++ src/main/java/com/example/market/dto/product/ProductResponseDto.java
@@ -20,7 +20,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class ProductResponseDto {
 
-	private final Long productId;
+	private final Integer productId;
 	private final String title;
 	private final Integer price;
 	private final ProductStatus status;
